requirements/base.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
sqlalchemy==2.0.23
alembic==1.12.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9  # For PostgreSQL support
python-dateutil==2.8.2
email-validator==2.1.0.post1
